Added More Functionalities
Added a lot of new functionalities such as save course attributes to a file, find course name from file and print a summary, print a list of all of the courses in the file, prevent saving dublicate course names to the file, added merge sort algorithm and vector, able to check if a course name already exist in the file and print a summary of that course. Made a seperate function for putting courses from file to array called FindEmptyIndex since the code is used in multiple places. Added an option for Find Course by name. Added a global variable find_course.